re-organize test files
